+++ Makie/test/SceneLike/makielayout.jl
@@ -534,6 +534,25 @@ end
     @test make_fig(linesegments!, rand(8)) isa Figure
 end
 
+@testset "Joint legend data gathering" begin
+    f = Figure()
+    ax1 = Axis(f[1, 1])
+    l1a = lines!(ax1, rand(10), label = "test a")
+    l1b = lines!(ax1, rand(10), label = "test b")
+    ax2 = Axis(f[1, 2])
+    l2a = lines!(ax2, rand(10), label = "test a")
+    leg = Legend(f[1, 3], [ax1, ax2], merge = true)
+
+    @test f isa Figure
+    # The joint legend has two entries
+    @test length(leg.entrygroups[][][2]) == 2
+    # The first entry has two linked plots
+    @test length(leg.entrygroups[][][2][1].elements) == 2
+    # The two linked plots are the plots from two different axes
+    @test leg.entrygroups[][][2][1].elements[1].plots[] == l1a
+    @test leg.entrygroups[][][2][1].elements[2].plots[] == l2a
+end
+
 @testset "ReversibleScale" begin
     @test ReversibleScale(identity).inverse === identity
     @test ReversibleScale(log).inverse === exp